apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bimatlas-api"
version = "0.1.0"
description = "BimAtlas -- FastAPI + Strawberry GraphQL backend for IFC spatial-graph engine"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "strawberry-graphql[fastapi]",
    "ifcopenshell",
    "psycopg2-binary",
    "apache-age-python",
    "numpy",
    "python-multipart",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "httpx",
    "ruff",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
